requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
python-jose
boto3
apscheduler
apache-airflow
pytest
